+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Notify authentication successes/failures for browser login and error paths in Client Script Based Authentication.
 
 ## [0.42.0] - 2026-08-26
 ### Added